requirements.txt
langgraph 
langchain-openai
tavily-python 
python-dotenv
streamlit
tenacity